requirements.txt
altgraph==0.17.2
cycler==0.11.0
fonttools==4.33.3
kiwisolver==1.4.3
matplotlib==3.5.2
MouseInfo==0.1.3
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.1.1
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyQt6==6.3.1
PyQt6-Qt6==6.3.1
PyQt6-sip==13.4.0
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.4
PyYAML==6.0
six==1.16.0
